[plugins/spotify] fix broken integration
I recommend not using purely hierarchical css/xpath selectors as they
will be very brittle, breaking on any minor change to the DOM tree.

Additionally, it's better to query/test for attributes/classnames that
aren't meant to contain text to be displayed to the user (e.g. don't use
titles as previously used in getPlaybackStatus), because it's likely
that Spotify's i18n will cause breakage.