Fix Spotify integration selectors
Fixes favorite button and art URL selectors.

Spotify has been changing their HTML a lot lately!

Now they've made it a bit more difficult to determine if a song is
fave'd or not. They're using Sass to generate CSS classes with random
names, nullifying the previous approach.

At the moment, the only ways I can think of to determine if the song is
fave'd:

a. test whether the fave button's title is "Save to Your Library" or "Remove from Your Library"
b. test whether the CSS color of the button is green (rgb(29, 185, 84)).
c. test the length of the button's classList, 1 for unfave'd and 2 for fave'd

Option *a* would fail if the user's language is not English, so that
one's off the table.

Option *b* would break if Spotify even slightly changes the color.

I've decided to go with Option *c*, which isn't ideal but seems like the
least brittle of the three.

(cherry picked from commit b75e6c161f04807a45601f0a6a817b1b8f059fa1)